+++ src/pages/CartPage/index.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import DocumentTitle from 'react-document-title';
 import PageHeader from '../../components/PageHeader';
-import { graphql, compose } from 'react-apollo';
-import gql from 'graphql-tag';
+import { compose } from 'react-apollo';
 import './CartPage.css';
 import CartItem from './CartItem';
 import SummaryRow from './SummaryRow';
